Rename: postFireStore.ts → postFirebaseStorage
